requirements.txt
fastapi==0.73.0
eth-brownie==1.16.0
requests==2.26.0
bs4==0.0.1
git+git://github.com/NurtureLabs/web3.py.git
Django==3.2.12
django-bulk-update-or-create==0.3.0
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-extensions==3.1.5
humanize==3.14.0
python-dotenv==0.16.0
uvicorn==0.17.4
pydantic==1.9.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
botocore==1.23.49
boto3==1.20.49
pycurl==7.44.1
celery==5.2.2
celery-pubsub==1.0.0
redis==3.5.3
sentry-sdk==1.5.4
pysha3==1.0.2
diskcache==5.4.0
slowapi==0.1.5
lxml==4.7.1
Pillow==9.0.1
pipe==1.6.0
flower==1.0.0
IPython
jupyter
jupyter_http_over_ws==0.0.8
autopep8==1.6.0
pandas==1.4.1
texttable==1.6.4
black
